changes while fixing "commands"
added or changed funcs

+++ project/Song.h
@@ -15,7 +15,7 @@ private:
 	std::string genre; 
 	std::string album;
 	size_t release_year = 0; 
-	std::pair <int, int> rating; //num of votes, sum of votes
+	std::pair <size_t, size_t> rating; //num of votes, sum of votes
 private:
 	//constructor helpers
 	void copy(const Song& other);
@@ -27,12 +27,12 @@ public:
 	//constructors
 	Song();
 	Song(const std::string& _name, const std::string& _artist, const std::string& _genre,
-		 const std::string& _album, int _release_year);
+		 const std::string& _album, size_t _release_year);
 	Song& operator= (const Song& other);
 	~Song();
 
 	//public or private? setters and getters
-	void set_rating(int rate);
+	void set_rating(size_t rate);
 	double get_rating() const;
 	std::string get_name() const;
